Extract risk attribution response module
