+++ src/app/services/risk_workspace_service.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from datetime import date, timedelta
 from typing import Any, cast
 
@@ -7,13 +6,7 @@ from fastapi import status
 from app.config import settings
 from app.contracts.risk_workspace import (
     RiskModuleState,
-    WorkbenchRiskAttributionContributor,
-    WorkbenchRiskAttributionControls,
-    WorkbenchRiskAttributionMethodologyContext,
-    WorkbenchRiskAttributionPayload,
-    WorkbenchRiskAttributionPeriodResult,
     WorkbenchRiskAttributionResponse,
-    WorkbenchRiskAttributionSet,
     WorkbenchRiskConcentrationResponse,
     WorkbenchRiskDrawdownResponse,
     WorkbenchRiskMetadata,
@@ -27,18 +20,12 @@ from app.contracts.risk_workspace import (
 from app.contracts.workbench import WorkbenchPartialFailure
 from app.services.async_ttl_cache import AsyncTtlCache
 from app.services.domain_client_protocols import RiskWorkspaceClient
-from app.services.risk_workspace_attribution_controls import (
-    RISK_ATTRIBUTION_ACTIVE_RISK_GATED_GROUPINGS as _RISK_ATTRIBUTION_ACTIVE_RISK_GATED_GROUPINGS,
-)
-from app.services.risk_workspace_attribution_controls import (
-    build_attribution_controls,
-    build_attribution_supportability,
-    metadata_grouping_dimension_set,
+from app.services.risk_workspace_attribution import (
+    blocked_attribution_response,
+    map_attribution_response,
     normalize_risk_attribution_grouping,
     normalize_risk_attribution_type,
-    resolve_active_risk_grouping_support,
-    risk_attribution_grouping_label,
-    total_risk_gated_grouping_reason,
+    unavailable_attribution,
 )
 from app.services.risk_workspace_concentration import (
     map_concentration_response,
@@ -89,13 +76,6 @@ _METRIC_LABELS = {
     "INFORMATION_RATIO": "Information Ratio",
     "VAR": "Value at Risk",
 }
-
-
-@dataclass(frozen=True)
-class AttributionMappingResult:
-    periods: list[WorkbenchRiskAttributionPeriodResult]
-    warnings: list[str]
-    partial_failures: list[WorkbenchPartialFailure]
 
 
 class RiskWorkspaceService:
@@ -462,7 +442,7 @@ class RiskWorkspaceService:
         resolved_as_of_date = _resolve_as_of_date(as_of_date)
         normalized_type = _normalize_risk_attribution_type(attribution_type)
         normalized_grouping = _normalize_risk_attribution_grouping(grouping_dimension)
-        blocked_response = _blocked_attribution_response(
+        blocked_response = blocked_attribution_response(
             correlation_id=correlation_id,
             portfolio_id=portfolio_id,
             period=period,
@@ -511,7 +491,7 @@ class RiskWorkspaceService:
             if upstream_status >= status.HTTP_400_BAD_REQUEST or not isinstance(
                 upstream_payload, dict
             ):
-                return _unavailable_attribution(
+                return unavailable_attribution(
                     correlation_id=correlation_id,
                     portfolio_id=portfolio_id,
                     period=period,
@@ -522,7 +502,7 @@ class RiskWorkspaceService:
                     upstream_status=upstream_status,
                     upstream_payload=upstream_payload,
                 )
-            return _map_attribution_response(
+            return map_attribution_response(
                 correlation_id=correlation_id,
                 portfolio_id=portfolio_id,
                 period=period,
@@ -711,241 +691,6 @@ def _metric_dependency_supportability(
     ]
 
 
-def _map_attribution_response(
-    *,
-    correlation_id: str,
-    portfolio_id: str,
-    period: str,
-    as_of_date: str,
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-    upstream_payload: dict[str, Any],
-) -> WorkbenchRiskAttributionResponse:
-    results = upstream_payload.get("results")
-    mapping = _map_attribution_period_results(
-        results=results,
-        attribution_type=attribution_type,
-        grouping_dimension=grouping_dimension,
-    )
-    supportability = _build_attribution_supportability(
-        benchmark_code=benchmark_code,
-        attribution_type=attribution_type,
-        grouping_dimension=grouping_dimension,
-    )
-    _append_source_calculation_supportability(
-        supportability=supportability,
-        upstream_payload=upstream_payload,
-    )
-    upstream_metadata = upstream_payload.get("metadata")
-    state, warnings, partial_failures = _resolve_attribution_state(
-        period_results=mapping.periods,
-        supportability=supportability,
-        warnings=mapping.warnings,
-        partial_failures=mapping.partial_failures,
-    )
-
-    return WorkbenchRiskAttributionResponse(
-        correlation_id=correlation_id,
-        portfolio_id=portfolio_id,
-        period=period,
-        as_of_date=as_of_date,
-        benchmark_code=benchmark_code,
-        state=state,
-        payload=_build_attribution_payload(
-            period_results=mapping.periods,
-            benchmark_code=benchmark_code,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-            upstream_metadata=upstream_metadata,
-        ),
-        supportability=supportability,
-        warnings=sorted(set(warnings)),
-        partial_failures=partial_failures,
-        metadata=_build_attribution_metadata(upstream_metadata=upstream_metadata),
-    )
-
-
-def _map_attribution_period_results(
-    *,
-    results: Any,
-    attribution_type: str,
-    grouping_dimension: str,
-) -> AttributionMappingResult:
-    warnings: list[str] = []
-    partial_failures: list[WorkbenchPartialFailure] = []
-    period_results: list[WorkbenchRiskAttributionPeriodResult] = []
-
-    if not isinstance(results, dict):
-        return AttributionMappingResult(
-            periods=period_results,
-            warnings=warnings,
-            partial_failures=partial_failures,
-        )
-
-    for key, value in results.items():
-        if not isinstance(value, dict):
-            continue
-        period = _map_attribution_period_result(
-            key=key,
-            value=value,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-        )
-        if period.error:
-            partial_failures.append(
-                WorkbenchPartialFailure(
-                    source_service="risk",
-                    error_code="RISK_ATTRIBUTION_PERIOD_ERROR",
-                    detail=f"{key}: {period.error}",
-                )
-            )
-            warnings.append("RISK_ATTRIBUTION_PERIOD_PARTIAL")
-        period_results.append(period)
-
-    return AttributionMappingResult(
-        periods=period_results,
-        warnings=warnings,
-        partial_failures=partial_failures,
-    )
-
-
-def _map_attribution_period_result(
-    *,
-    key: Any,
-    value: dict[str, Any],
-    attribution_type: str,
-    grouping_dimension: str,
-) -> WorkbenchRiskAttributionPeriodResult:
-    error = value.get("error")
-    return WorkbenchRiskAttributionPeriodResult(
-        key=str(key),
-        label=str(key),
-        start_date=str(value.get("start_date", "")),
-        end_date=str(value.get("end_date", "")),
-        attribution_sets=_map_attribution_sets(
-            value.get("attribution_sets"),
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-        ),
-        error=str(error) if isinstance(error, str) and error.strip() else None,
-    )
-
-
-def _map_attribution_sets(
-    value: Any,
-    *,
-    attribution_type: str,
-    grouping_dimension: str,
-) -> list[WorkbenchRiskAttributionSet]:
-    if not isinstance(value, list):
-        return []
-    return [
-        _map_attribution_set(
-            entry,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-        )
-        for entry in value
-        if isinstance(entry, dict)
-    ]
-
-
-def _map_attribution_set(
-    entry: dict[str, Any],
-    *,
-    attribution_type: str,
-    grouping_dimension: str,
-) -> WorkbenchRiskAttributionSet:
-    return WorkbenchRiskAttributionSet(
-        attribution_type=str(entry.get("attribution_type", attribution_type)),
-        metric=str(entry.get("metric", "")),
-        grouping_dimension=str(entry.get("grouping_dimension", grouping_dimension)),
-        total_value=_safe_float(entry.get("total_value")),
-        reconciled_sum=_safe_float(entry.get("reconciled_sum")),
-        residual=_safe_float(entry.get("residual")),
-        contributors=_map_attribution_contributors(entry.get("contributors")),
-        quality_flags=[
-            str(flag)
-            for flag in entry.get("quality_flags", [])
-            if isinstance(flag, str) and flag.strip()
-        ],
-    )
-
-
-def _map_attribution_contributors(value: Any) -> list[WorkbenchRiskAttributionContributor]:
-    if not isinstance(value, list):
-        return []
-    return [
-        WorkbenchRiskAttributionContributor.model_validate(contributor)
-        for contributor in value
-        if isinstance(contributor, dict)
-    ]
-
-
-def _resolve_attribution_state(
-    *,
-    period_results: list[WorkbenchRiskAttributionPeriodResult],
-    supportability: list[WorkbenchRiskSupportabilityItem],
-    warnings: list[str],
-    partial_failures: list[WorkbenchPartialFailure],
-) -> tuple[RiskModuleState, list[str], list[WorkbenchPartialFailure]]:
-    resolved_warnings = list(warnings)
-    resolved_partial_failures = list(partial_failures)
-    state: RiskModuleState = (
-        "partial" if any(item.state != "ready" for item in supportability) else "ready"
-    )
-    if not period_results:
-        state = "unavailable"
-        resolved_warnings.append("RISK_ATTRIBUTION_EMPTY")
-        resolved_partial_failures.append(
-            WorkbenchPartialFailure(
-                source_service="risk",
-                error_code="EMPTY_RISK_ATTRIBUTION",
-                detail="lotus-risk returned no attribution periods.",
-            )
-        )
-    elif all(not period.attribution_sets for period in period_results):
-        state = "unavailable"
-    return state, resolved_warnings, resolved_partial_failures
-
-
-def _build_attribution_metadata(*, upstream_metadata: Any) -> WorkbenchRiskMetadata:
-    metadata = _metadata(input_mode="stateful", cache_status="miss")
-    if isinstance(upstream_metadata, dict):
-        methodology_version = upstream_metadata.get("methodology_version")
-        if isinstance(methodology_version, str) and methodology_version.strip():
-            return metadata.model_copy(update={"methodology_version": methodology_version})
-    return metadata
-
-
-def _build_attribution_payload(
-    *,
-    period_results: list[WorkbenchRiskAttributionPeriodResult],
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-    upstream_metadata: Any,
-) -> WorkbenchRiskAttributionPayload | None:
-    if not period_results:
-        return None
-    metadata = upstream_metadata if isinstance(upstream_metadata, dict) else None
-    return WorkbenchRiskAttributionPayload(
-        controls=_build_attribution_controls(
-            benchmark_code=benchmark_code,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-            upstream_metadata=metadata,
-        ),
-        periods=period_results,
-        methodology_context=(
-            WorkbenchRiskAttributionMethodologyContext.model_validate(upstream_metadata)
-            if isinstance(upstream_metadata, dict)
-            else None
-        ),
-    )
-
-
 def _append_source_calculation_supportability(
     *,
     supportability: list[WorkbenchRiskSupportabilityItem],
@@ -967,86 +712,6 @@ def _append_source_calculation_supportability(
             source_service=source_supportability.source_service or "lotus-risk",
         )
     )
-
-
-def _build_attribution_controls(
-    *,
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-    upstream_metadata: dict[str, Any] | None = None,
-) -> WorkbenchRiskAttributionControls:
-    return build_attribution_controls(
-        benchmark_code=benchmark_code,
-        attribution_type=attribution_type,
-        grouping_dimension=grouping_dimension,
-        upstream_metadata=upstream_metadata,
-    )
-
-
-def _build_attribution_supportability(
-    *,
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-    upstream_metadata: dict[str, Any] | None = None,
-) -> list[WorkbenchRiskSupportabilityItem]:
-    return build_attribution_supportability(
-        benchmark_code=benchmark_code,
-        attribution_type=attribution_type,
-        grouping_dimension=grouping_dimension,
-        upstream_metadata=upstream_metadata,
-    )
-
-
-def _blocked_attribution_response(
-    *,
-    correlation_id: str,
-    portfolio_id: str,
-    period: str,
-    as_of_date: str,
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-) -> WorkbenchRiskAttributionResponse | None:
-    is_active_risk_without_benchmark = attribution_type == "ACTIVE_RISK" and not benchmark_code
-    is_active_risk_gated_grouping = (
-        attribution_type == "ACTIVE_RISK"
-        and grouping_dimension in _RISK_ATTRIBUTION_ACTIVE_RISK_GATED_GROUPINGS
-    )
-    if not is_active_risk_without_benchmark and not is_active_risk_gated_grouping:
-        return None
-    controls = _build_attribution_controls(
-        benchmark_code=benchmark_code,
-        attribution_type=attribution_type,
-        grouping_dimension=grouping_dimension,
-    )
-    return WorkbenchRiskAttributionResponse(
-        correlation_id=correlation_id,
-        portfolio_id=portfolio_id,
-        period=period,
-        as_of_date=as_of_date,
-        benchmark_code=benchmark_code,
-        state="blocked",
-        payload=WorkbenchRiskAttributionPayload(controls=controls, periods=[]),
-        supportability=_build_attribution_supportability(
-            benchmark_code=benchmark_code,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-        ),
-        warnings=["RISK_ATTRIBUTION_BLOCKED"],
-        partial_failures=[],
-        metadata=_metadata(input_mode="stateful", cache_status="bypass"),
-    )
-
-
-def _safe_float(value: Any) -> float | None:
-    if value is None:
-        return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return None
 
 
 def _unavailable_summary(
@@ -1071,49 +736,6 @@ def _unavailable_summary(
             reason="lotus-risk summary endpoint is unavailable."
         ),
         warnings=["RISK_SUMMARY_UNAVAILABLE"],
-        partial_failures=[
-            risk_upstream_failure(
-                upstream_status=upstream_status,
-                upstream_payload=upstream_payload,
-            )
-        ],
-        metadata=_metadata(input_mode="stateful", cache_status="miss"),
-    )
-
-
-def _unavailable_attribution(
-    *,
-    correlation_id: str,
-    portfolio_id: str,
-    period: str,
-    as_of_date: str,
-    benchmark_code: str | None,
-    attribution_type: str,
-    grouping_dimension: str,
-    upstream_status: int,
-    upstream_payload: Any,
-) -> WorkbenchRiskAttributionResponse:
-    return WorkbenchRiskAttributionResponse(
-        correlation_id=correlation_id,
-        portfolio_id=portfolio_id,
-        period=period,
-        as_of_date=as_of_date,
-        benchmark_code=benchmark_code,
-        state="unavailable",
-        payload=WorkbenchRiskAttributionPayload(
-            controls=_build_attribution_controls(
-                benchmark_code=benchmark_code,
-                attribution_type=attribution_type,
-                grouping_dimension=grouping_dimension,
-            ),
-            periods=[],
-        ),
-        supportability=_build_attribution_supportability(
-            benchmark_code=benchmark_code,
-            attribution_type=attribution_type,
-            grouping_dimension=grouping_dimension,
-        ),
-        warnings=["RISK_ATTRIBUTION_UNAVAILABLE"],
         partial_failures=[
             risk_upstream_failure(
                 upstream_status=upstream_status,
@@ -1155,31 +777,4 @@ def _build_risk_periods(
         period=period,
         report_start_date=report_start_date,
         report_end_date=report_end_date,
-    )
-
-
-def _risk_attribution_grouping_label(grouping_key: str) -> str:
-    return risk_attribution_grouping_label(grouping_key)
-
-
-def _resolve_active_risk_grouping_support(
-    metadata: dict[str, Any] | None,
-) -> tuple[set[str], set[str], str]:
-    return resolve_active_risk_grouping_support(metadata)
-
-
-def _total_risk_gated_grouping_reason(active_risk_gate_reason: str | None) -> str:
-    return total_risk_gated_grouping_reason(active_risk_gate_reason)
-
-
-def _metadata_grouping_dimension_set(
-    *,
-    metadata: dict[str, Any],
-    field_name: str,
-    default: tuple[str, ...],
-) -> set[str]:
-    return metadata_grouping_dimension_set(
-        metadata=metadata,
-        field_name=field_name,
-        default=default,
     )
